requirements.txt
pygame==2.5.2
openai
python-dotenv
